pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-site-proxy"
version = "0.1.0"
description = "Local FastAPI proxy and discovery workflow for site APIs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110,<1.0",
  "httpx>=0.27,<1.0",
  "playwright>=1.41,<2.0",
  "pydantic>=1.10,<3.0",
  "PyYAML>=6.0,<7.0",
  "uvicorn>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
]

[project.scripts]
usp = "universal_site_proxy.cli:main"

[tool.pytest.ini_options]
markers = [
  "playwright: tests that require a local Playwright browser installation",
]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["universal_site_proxy*", "discovery*"]

[tool.setuptools.package-data]
universal_site_proxy = ["api_maps/*.yaml"]
